etermax_api_service/requirements/local.txt
celery==5.4.0
Django==5.0.4
djangorestframework==3.15.1
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-redis==5.4.0
pytest==8.2.1
python-dotenv==1.0.1
requests==2.32.3

pdbp==1.5.0
